Remove manifest.sepolicy.version from device manifest
Fixes:
FAILED: out/target/product/ibiza/gen/ETC/vendor_manifest.xml_intermediates/manifest.xml
/bin/bash -c "BOARD_SEPOLICY_VERS=32.0 	PRODUCT_ENFORCE_VINTF_MANIFEST=false 	PRODUCT_SHIPPING_API_LEVEL= 	out/host/linux-x86/bin/assemble_vintf -o out/target/product/ibiza/gen/ETC/vendor_manifest.xml_intermediates/manifest.xml 		-i device/motorola/ibiza/manifest.xml"
Cannot override existing value 30.0 with BOARD_SEPOLICY_VERS (which is 32.0).

ibiza shipped with Android 11 (API 30), we're building Android 12L/12.1
(API 32), so that's where the default value and manifest value is coming
from. If we instead set BOARD_SEPOLICY_VERS to 32 it brings in a
whole bunch of sepolicy stuff that we want to avoid for now (and this
property will apparently be added automatically)